Formal Version 1.0

+++ gp_cross_factor/utils.py
@@ -1,4 +1,5 @@
-"""Utilities that are required by gplearn.
+"""
+Utilities that are required by gplearn.
 
 Most of these functions are slightly modified versions of some key utility
 functions from scikit-learn that gplearn depends upon. They reside here in
@@ -68,8 +69,7 @@ def _partition_estimators(n_estimators, n_jobs):
     n_jobs = min(_get_n_jobs(n_jobs), n_estimators)
 
     # Partition estimators between jobs
-    n_estimators_per_job = (n_estimators // n_jobs) * np.ones(n_jobs,
-                                                              dtype=int)
+    n_estimators_per_job = (n_estimators // n_jobs) * np.ones(n_jobs, dtype=int)
     n_estimators_per_job[:n_estimators % n_jobs] += 1
     starts = np.cumsum(n_estimators_per_job)
